server/requirements.txt
aiofiles==0.6.0
aiosqlite==0.16.1
asyncpg==0.22.0
certifi==2020.12.5
chardet==4.0.0
gevent==21.1.2
greenlet==1.0.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httpcore==0.12.3
httptools==0.1.1
httpx==0.16.1
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
iso8601==0.1.14
multidict==5.1.0
paho-mqtt==1.5.1
priority==1.3.0
pypika-tortoise==0.1.0
python-dotenv==0.17.0
pytz==2020.5
requests==2.25.1
rfc3986==1.4.0
sanic==21.3.2
sanic-routing==0.5.2
sanic-testing==0.3.0
sniffio==1.2.0
toml==0.10.2
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
uvloop==0.15.2
websockets==8.1
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.3.0
